fix: 修复合并/清理日志中磁盘空间变化显示错误
uint64 减法在磁盘占用增加时下溢为巨大正数（如 17179869181.2 GB），
转为 int64 减法避免下溢。同时补充 clean 日志中磁盘占用增加时的显示。

+++ internal/service/merge.go
@@ -328,7 +328,7 @@ func (s *MergeService) summarizeResults(ctx context.Context, root, streamer stri
 
 	// 磁盘变化
 	if diskStart != nil && diskEnd != nil {
-		freeChangeGB := float64(diskEnd.Free-diskStart.Free) / oneGB
+		freeChangeGB := float64(int64(diskEnd.Free)-int64(diskStart.Free)) / oneGB
 		if freeChangeGB > 0 {
 			onProgress(fmt.Sprintf("  磁盘：%.1f%% → %.1f%%（释放 %.1f GB）", diskStart.UsedPct, diskEnd.UsedPct, freeChangeGB))
 		} else {
